+++ doc/release-notes.md
@@ -8,4 +8,7 @@ This release includes the following features and fixes:
  - The Chronik Electrum blockchain.block.headers RPC now returns up to
    2016 block headers, as per the documentation. Previously, it mistakenly
    applied the page limits relevant to Chronik's block and headers endpoints
-   and thus return an error if more than 500 headers were requested.
+   and thus returned an error if more than 500 headers were requested.
+ - The getblockchaininfo RPC now has a new field "finalized_blockhash" which
+   gives the hash of the highest block that is finalized by Avalanche, or
+   the hash of the genesis block if no block is finalized by Avalanche.